daemon: refactor job api

+++ packages/daemon/src/service/pipeline.ts
@@ -4,6 +4,7 @@ import * as fs from 'fs-extra';
 import { generate } from 'shortid';
 import { Op } from 'sequelize';
 import { provide, inject, Context } from 'midway';
+import * as HttpStatus from 'http-status';
 import {
   PipelineDB,
   PipelineStatus,
@@ -18,6 +19,7 @@ import { PluginPackage, RunnableResponse, PluginRunnable } from '@pipcook/costa'
 import { PipelineModel, PipelineModelStatic } from '../model/pipeline';
 import { JobModelStatic, JobModel } from '../model/job';
 import { PluginManager } from './plugin';
+import { LogObject } from './log-manager';
 
 interface QueryOptions {
   limit: number;
@@ -57,7 +59,7 @@ export class PipelineService {
 
   @inject()
   ctx: Context;
-  
+
   @inject('pipelineModel')
   pipeline: PipelineModelStatic;
 
@@ -123,29 +125,39 @@ export class PipelineService {
     });
   }
 
-  async queryJobs(filter: SelectJobsFilter, opts?: QueryOptions): Promise<{rows: JobModel[], count: number}> {
+  async queryJobs(filter: SelectJobsFilter, opts?: QueryOptions): Promise<JobModel[]> {
     const where = {} as any;
     const { offset, limit } = opts || {};
     if (typeof filter.pipelineId === 'string') {
       where.pipelineId = filter.pipelineId;
     }
-    return this.job.findAndCountAll({
+    return (await this.job.findAndCountAll({
       offset,
       limit,
       where,
       order: [
         [ 'createdAt', 'DESC' ]
       ]
-    });
+    })).rows;
   }
 
   async removeJobs(): Promise<number> {
     const jobs = await this.queryJobs({});
-    await jobs.rows.map(async (job: JobModel) => {
+    await jobs.map(async (job: JobModel) => {
+      await job.destroy();
+      fs.remove(`${CoreConstants.PIPCOOK_RUN}/${job.id}`);
+    });
+    return jobs.length;
+  }
+
+  async removeJobById(id: string): Promise<number> {
+    const job = await this.job.findByPk(id);
+    if (job) {
       await job.destroy();
       await fs.remove(`${CoreConstants.PIPCOOK_RUN}/${job.id}`);
-    });
-    return jobs.rows.length;
+      return 1;
+    }
+    return 0;
   }
 
   async createJob(pipelineId: string): Promise<JobModel> {
@@ -160,13 +172,12 @@ export class PipelineService {
     return job;
   }
 
-  async installPlugins(job: JobModel, pyIndex?: string): Promise<Partial<Record<PluginTypeI, PluginInfo>>> {
-    const pipeline = await this.getPipeline(job.pipelineId);
+  async installPlugins(pipeline: PipelineModel, log: LogObject, pyIndex?: string): Promise<Partial<Record<PluginTypeI, PluginInfo>>> {
     const plugins: Partial<Record<PluginTypeI, PluginInfo>> = {};
     for (const type of CoreConstants.PLUGINS) {
       if (pipeline[type]) {
         plugins[type] = await {
-          plugin: await this.pluginManager.fetchAndInstall(pipeline[type], pyIndex),
+          plugin: await this.pluginManager.fetchAndInstall(pipeline[type], log, pyIndex),
           params: pipeline[`${type}Params`]
         };
       }
@@ -174,8 +185,8 @@ export class PipelineService {
     return plugins;
   }
 
-  async startJob(job: JobModel, plugins: Partial<Record<PluginTypeI, PluginInfo>>) {
-    const runnable = await this.pluginManager.createRunnable(job.id);
+  async startJob(job: JobModel, pipeline: PipelineModel, plugins: Partial<Record<PluginTypeI, PluginInfo>>, logger: LogObject): Promise<void> {
+    const runnable = await this.pluginManager.createRunnable(job.id, logger);
     // save the runnable object
     runnableMap[job.id] = runnable;
 
@@ -247,17 +258,6 @@ export class PipelineService {
         modelDir: modelPath
       }));
 
-      const result = await runnable.valueOf(output) as EvaluateResult;
-      job.evaluateMap = JSON.stringify(result);
-      job.evaluatePass = result.pass;
-      job.endTime = Date.now();
-      job.status = PipelineStatus.SUCCESS;
-      const datasetVal = await runnable.valueOf(dataset) as UniDataset;
-      if (datasetVal?.metadata) {
-        job.dataset = JSON.stringify(datasetVal.metadata);
-      }
-      await job.save();
-      const pipeline = await this.getPipeline(job.pipelineId);
       await this.generateOutput(job, {
         modelPath,
         modelPlugin,
@@ -266,6 +266,18 @@ export class PipelineService {
         workingDir: runnable.workingDir,
         template: 'node' // set node by default
       });
+
+      // update job status to successful
+      const result = await runnable.valueOf(output) as EvaluateResult;
+      const datasetVal = await runnable.valueOf(dataset) as UniDataset;
+      if (datasetVal?.metadata) {
+        job.dataset = JSON.stringify(datasetVal.metadata);
+      }
+      job.evaluateMap = JSON.stringify(result);
+      job.evaluatePass = result.pass;
+      job.endTime = Date.now();
+      job.status = PipelineStatus.SUCCESS;
+      await job.save();
     } catch (err) {
       job.status = PipelineStatus.FAIL;
       job.error = err.message;
@@ -283,7 +295,7 @@ export class PipelineService {
       runnable.destroy();
       delete runnableMap[id];
     } else {
-      this.ctx.throw('no runnable found');
+      this.ctx.throw(HttpStatus.NOT_FOUND, 'no runnable found');
     }
   }
 
@@ -350,5 +362,10 @@ export class PipelineService {
       await fs.readFile(stdout, 'utf8'),
       await fs.readFile(stderr, 'utf8')
     ];
+  }
+
+  async runJob(job: JobModel, pipeline: PipelineModel, log: LogObject): Promise<void> {
+    const plugins = await this.installPlugins(pipeline, log);
+    await this.startJob(job, pipeline, plugins, log);
   }
 }
